+++ TeamSphere-backend/src/test/com/teamsphere/service/impl/ProjectServiceImplTest.java
@@ -1,0 +1,116 @@
+package com.teamsphere.service.impl;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+import com.teamsphere.dto.project.ProjectDto;
+import com.teamsphere.dto.project.ProjectSearchRequest;
+import com.teamsphere.entity.ProjectEntity;
+import com.teamsphere.entity.enums.ProjectStatus;
+import com.teamsphere.mapper.ProjectMapper;
+import com.teamsphere.repository.ProjectRepository;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.TypedQuery;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+
+import java.time.LocalDate;
+import java.util.Collections;
+
+@ExtendWith(MockitoExtension.class)
+class ProjectServiceImplTest {
+
+    @Mock
+    private ProjectRepository projectRepository;
+
+    @Mock
+    private ProjectMapper projectMapper;
+
+    @Mock
+    private EntityManager entityManager;
+
+    @InjectMocks
+    private ProjectServiceImpl projectService;
+
+    private ProjectDto projectDto;
+    private ProjectEntity projectEntity;
+
+    @BeforeEach
+    void setUp() {
+        projectDto = ProjectDto.builder()
+                .id(1L)
+                .name("Test Project")
+                .description("Test Description")
+                .status(ProjectStatus.IN_PROGRESS.toString())
+                .startDate(LocalDate.now().toString())
+                .build();
+
+        projectEntity = new ProjectEntity();
+        projectEntity.setId(1L);
+        projectEntity.setName("Test Project");
+        projectEntity.setDescription("Test Description");
+        projectEntity.setStatus(ProjectStatus.IN_PROGRESS);
+        projectEntity.setStartDate(LocalDate.now());
+    }
+
+    @Test
+    void testGetAll() {
+        Pageable pageable = PageRequest.of(0, 10);
+        when(projectRepository.findAllWithCompanies()).thenReturn(Collections.singletonList(projectEntity));
+        when(projectMapper.toDto(any(ProjectEntity.class))).thenReturn(projectDto);
+
+        Page<ProjectDto> result = projectService.getAll(pageable);
+
+        assertEquals(1, result.getTotalElements());
+        assertEquals(projectDto, result.getContent().getFirst());
+    }
+
+    @Test
+    void testFind() {
+        ProjectSearchRequest request = new ProjectSearchRequest("Test");
+        Pageable pageable = PageRequest.of(0, 10);
+
+        CriteriaBuilder criteriaBuilder = mock(CriteriaBuilder.class);
+        CriteriaQuery<ProjectEntity> criteriaQuery = mock(CriteriaQuery.class);
+        Root<ProjectEntity> root = mock(Root.class);
+        Predicate predicate = mock(Predicate.class);
+        TypedQuery<ProjectEntity> typedQuery = mock(TypedQuery.class);
+        CriteriaQuery<Long> countQuery = mock(CriteriaQuery.class);
+        TypedQuery<Long> countTypedQuery = mock(TypedQuery.class);
+
+        when(entityManager.getCriteriaBuilder()).thenReturn(criteriaBuilder);
+        when(criteriaBuilder.createQuery(ProjectEntity.class)).thenReturn(criteriaQuery);
+        when(criteriaQuery.from(ProjectEntity.class)).thenReturn(root);
+        when(criteriaBuilder.like(any(), any(String.class))).thenReturn(predicate);
+        when(criteriaBuilder.or(any(Predicate.class), any(Predicate.class))).thenReturn(predicate);
+        when(entityManager.createQuery(criteriaQuery)).thenReturn(typedQuery);
+        when(typedQuery.setFirstResult(any(int.class))).thenReturn(typedQuery);
+        when(typedQuery.setMaxResults(any(int.class))).thenReturn(typedQuery);
+        when(typedQuery.getResultList()).thenReturn(Collections.singletonList(projectEntity));
+        when(projectMapper.toDto(any(ProjectEntity.class))).thenReturn(projectDto);
+
+        when(criteriaBuilder.createQuery(Long.class)).thenReturn(countQuery);
+        when(countQuery.from(any(Class.class))).thenReturn(mock(Root.class));
+        when(countQuery.select(any())).thenReturn(countQuery);
+        when(countQuery.where(any(Predicate.class))).thenReturn(countQuery);
+        when(entityManager.createQuery(countQuery)).thenReturn(countTypedQuery);
+        when(countTypedQuery.getSingleResult()).thenReturn(1L);
+
+        Page<ProjectDto> result = projectService.find(request, pageable);
+
+        assertEquals(1, result.getTotalElements());
+        assertEquals(projectDto, result.getContent().getFirst());
+    }
+}
