--- TeamSphere-backend/test/java/com/api/management/user/TeamSphereApplicationTests.java
@@ -1,14 +1,0 @@
-package com.api.management.user;
-
-import com.teamsphere.TeamSphereApplication;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = TeamSphereApplication.class)
-class TeamSphereApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
