--- TeamSphere-backend/test/java/com/teamsphere/config/JwtAuthenticationFilterTest.java
@@ -1,116 +1,0 @@
-package com.teamsphere.config;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
-
-@ExtendWith(MockitoExtension.class)
-public class JwtAuthenticationFilterTest {
-
-    @Mock
-    private JwtService jwtService;
-
-    @Mock
-    private UserDetailsService userDetailsService;
-
-    @Mock
-    private HttpServletRequest request;
-
-    @Mock
-    private HttpServletResponse response;
-
-    @Mock
-    private FilterChain filterChain;
-
-    @InjectMocks
-    private JwtAuthenticationFilter jwtAuthenticationFilter;
-
-    @BeforeEach
-    void setUp() {
-        SecurityContextHolder.clearContext();
-    }
-
-    @Test
-    void doFilterInternal_shouldAuthenticateUserWithValidToken() throws ServletException, IOException {
-        // Given
-        String jwt = "valid-jwt";
-        String userEmail = "test@example.com";
-        UserDetails userDetails = new User(userEmail, "password", Collections.emptyList());
-
-        when(request.getHeader("Authorization")).thenReturn("Bearer " + jwt);
-        when(jwtService.extractUsername(jwt)).thenReturn(userEmail);
-        when(userDetailsService.loadUserByUsername(userEmail)).thenReturn(userDetails);
-        when(jwtService.isTokenValid(jwt, userDetails)).thenReturn(true);
-
-        // When
-        jwtAuthenticationFilter.doFilterInternal(request, response, filterChain);
-
-        // Then
-        verify(filterChain).doFilter(request, response);
-        assertNotNull(SecurityContextHolder.getContext().getAuthentication());
-        assertEquals(userEmail, ((UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername());
-    }
-
-    @Test
-    void doFilterInternal_shouldNotAuthenticateWithInvalidToken() throws ServletException, IOException {
-        // Given
-        String jwt = "invalid-jwt";
-        String userEmail = "test@example.com";
-        UserDetails userDetails = new User(userEmail, "password", Collections.emptyList());
-
-        when(request.getHeader("Authorization")).thenReturn("Bearer " + jwt);
-        when(jwtService.extractUsername(jwt)).thenReturn(userEmail);
-        when(userDetailsService.loadUserByUsername(userEmail)).thenReturn(userDetails);
-        when(jwtService.isTokenValid(jwt, userDetails)).thenReturn(false);
-
-        // When
-        jwtAuthenticationFilter.doFilterInternal(request, response, filterChain);
-
-        // Then
-        verify(filterChain).doFilter(request, response);
-        assertNull(SecurityContextHolder.getContext().getAuthentication());
-    }
-
-    @Test
-    void doFilterInternal_shouldNotAuthenticateWithMissingHeader() throws ServletException, IOException {
-        // Given
-        when(request.getHeader("Authorization")).thenReturn(null);
-
-        // When
-        jwtAuthenticationFilter.doFilterInternal(request, response, filterChain);
-
-        // Then
-        verify(filterChain).doFilter(request, response);
-        assertNull(SecurityContextHolder.getContext().getAuthentication());
-    }
-
-    @Test
-    void doFilterInternal_shouldNotAuthenticateWithInvalidHeader() throws ServletException, IOException {
-        // Given
-        when(request.getHeader("Authorization")).thenReturn("Invalid Header");
-
-        // When
-        jwtAuthenticationFilter.doFilterInternal(request, response, filterChain);
-
-        // Then
-        verify(filterChain).doFilter(request, response);
-        assertNull(SecurityContextHolder.getContext().getAuthentication());
-    }
-}
