--- TeamSphere-backend/test/resources/application.properties
@@ -1,9 +1,0 @@
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.defer-datasource-initialization=true
-jwt.secret=test-secret-key-for-testing-purposes-only-and-should-be-long-enough
-spring.sql.init.mode=never
